remove obsolete class `TryCatchBlocksAssertion`
... since it adds no value beyond defining the nested class `TryCatchBlockCondition`.
